Add execise solutions
For the loops and functions, I mostly just turned the instruction
on the page [1] into my algorithm. And for the next part, I
just wrote a function that would check if the first 4 decimal
places of temp hasn't changed. If so, it's the square root.

For the errors, again, it's mostly on the page translated into
code. On line 15 however instead of using the I get an error
if I write ErrNegativeSqrt(x).Error()

[1] https://tour.golang.org/flowcontrol/8
[2] https://tour.golang.org/methods/20