Add rot13reader solution
https://tour.golang.org/methods/23

First, read run the Read method on rot13Reader.r then apply
each byte to the rot13 helper method and return the n and
err.

The rot13 helper method is pretty straightforward. First, check
if it's a capital letter. Then check if it's not a letter. If not,
then return it right away. Then if it's less than 'N', add 13 to
it. Otherwise, minus 13.